update to errorHandling and save creds aftr prompt
