+++ src/job/ZosJobsProvider.ts
@@ -695,9 +695,7 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
                 // session = await getValidSession(profileLoaded, profileLoaded.name, false);
                 session = await ZoweExplorerApiRegister.getJesApi(profileLoaded).getSession();
             } catch (error) {
-                // When no password is entered, we should silence the error message for not providing it
-                // since password is optional in Zowe Explorer
-                if (error.message !== "Must have user & password OR base64 encoded credentials") { await errorHandling(error); }
+                await errorHandling(error);
             }
             // Creates ZoweNode to track new session and pushes it to mSessionNodes
             const node = new Job(profileLoaded.name, vscode.TreeItemCollapsibleState.Collapsed, null, session, null, profileLoaded);
