+++ src/utils.ts
@@ -20,8 +20,7 @@ import { IZoweTreeNode } from "./api/IZoweTreeNode";
 import * as nls from "vscode-nls";
 import { Profiles } from "./Profiles";
 // import { getValidSession } from "./profiles/utils";
-import { ZoweExplorerApiRegister } from "./api/ZoweExplorerApiRegister";
-import { promptCredentials } from "./profiles/utils";
+
 
 // Set up localization
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
@@ -37,6 +36,11 @@ export async function errorHandling(errorDetails: any, label?: string, moreInfo?
     let httpErrCode = null;
     const errMsg = localize("errorHandling.invalid.credentials", "Invalid Credentials. Please ensure the token or username/password for {0} are valid or this may lead to a lock-out.", label);
 
+    // When no password is entered, we should silence the error message for not providing it
+    // since password is optional in Zowe Explorer
+    if (errorDetails.message === "Must have user & password OR base64 encoded credentials") {
+        return;
+    }
     if (errorDetails.mDetails !== undefined) {
         httpErrCode = errorDetails.mDetails.errorCode;
     }
@@ -59,17 +63,15 @@ export async function errorHandling(errorDetails: any, label?: string, moreInfo?
                     // await getValidSession(invalidProfile, invalidProfile.name, true);
                     // await ZoweExplorerApiRegister.getUssApi(invalidProfile).getSession();
                     // await Profiles.getInstance().promptCredentials(label.trim(), true);
-                    promptCredentials(invalidProfile);
+                    Profiles.getInstance().promptCredentials(invalidProfile, true);
                 } else {
                     vscode.window.showErrorMessage(errMsg, "Check Credentials").then(async (selection) => {
                         if (selection) {
                             delete invalidProfile.profile.user;
                             delete invalidProfile.profile.password;
-                            // tslint:disable-next-line:no-console
-                            console.log(invalidProfile);
                             // await getValidSession(invalidProfile, invalidProfile.name, true);
                             // await ZoweExplorerApiRegister.getUssApi(invalidProfile).getSession();
-                            promptCredentials(invalidProfile);
+                            Profiles.getInstance().promptCredentials(invalidProfile, true);
                         }
                     });
                 }
